+++ src/main/java/com/example/eggcakeshopapi/service/ProductService.java
@@ -6,12 +6,12 @@ import com.example.eggcakeshopapi.repository.ProductRequest;
 import java.util.List;
 
 public interface ProductService {
-    //  查詢Product(所有)
+    //  Read查詢Product(所有)
     List<Product> getAllProducts();
-    //  查詢Product=>ID
+    //  Read查詢Product=>ID
     Product getProductById(Long productId) ;
-    //   新增Product=>ID
-    Integer createProduct(ProductRequest productRequest);
+    //   creat新增Product=>ID
+    Long createProduct(ProductRequest productRequest);
     //   更新Product=>ID
     void updateProduct(Long productId,ProductRequest productRequest);
     //   刪除Product=>ID
